refactor article fetch into cms container
