+++ frontend/src/App.js
@@ -24,13 +24,10 @@ class App extends Component {
   /* eslint-disable react/sort-comp */
   guestUser = { _id: '0', username: 'Guest' };
   state = {
-    articles: [],
-    articleIndex: {},
     startup: true,
     isLoggedIn: false,
     waitingForLogin: false,
     user: this.guestUser,
-    cmsReady: false,
   };
   componentDidMount = () => {
     if (this.state.waitingForLogin) {
@@ -66,33 +63,6 @@ class App extends Component {
           user: this.guestUser,
         });
       });
-
-    /*  bypass local storage
-
-    actions.checkLocalStorage().then(loaded => {
-      this.setState({
-        cmsReady: loaded,
-        articles: loaded ? JSON.parse(localStorage.getItem('allArticles')) : [],
-        articleIndex: loaded
-          ? JSON.parse(localStorage.getItem('articleIndex'))
-          : {},
-      });
-    });
-
-   */
-
-    if (this.state.articles.length === 0) {
-      actions
-        .getArticlesJSONBypass()
-        .then(res => {
-          this.setState({ articles: res, cmsReady: true });
-        })
-        .catch(err => {
-          // eslint-disable-next-line
-          console.log(err);
-          this.setState({ cmsReady: false });
-        });
-    }
   };
 
   setUser = user => {
@@ -116,9 +86,6 @@ class App extends Component {
 
   render() {
     const {
-      articles,
-      articleIndex,
-      cmsReady,
       isLoggedIn,
       startup,
       waitingForLogin,
@@ -186,29 +153,7 @@ class App extends Component {
               path="/assets/:id"
               render={props => <AssetEdit id={props.match.params.id} />}
             />
-            <Route
-              exact
-              path="/cms"
-              render={props => (
-                <CMSContainer
-                  {...props}
-                  articles={articles}
-                  articleIndex={articleIndex}
-                  cmsReady={cmsReady}
-                />
-              )}
-            />
-            <Route
-              path="/cms/:articleId"
-              render={props => (
-                <CMSContainer
-                  {...props}
-                  articles={articles}
-                  articleIndex={articleIndex}
-                  cmsReady={cmsReady}
-                />
-              )}
-            />
+            <Route path="/cms" component={CMSContainer} />
             <Route
               path="/login"
               render={() => <Login setUser={this.setUser} />}
